feat: implement NHTSA Api

+++ library/src/commonMain/kotlin/io/github/kabirnayeem99/viminfo/network/dto/NhtsaDecodeVinDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class NhtsaDecodeVinDto(
-    @SerialName("Count") val count: Int? = null,
+    @SerialName("Count") val count: Long? = null,
     @SerialName("Message") val message: String? = null,
     @SerialName("Results") val results: List<Result?>? = null,
     @SerialName("SearchCriteria") val searchCriteria: String? = null
@@ -16,14 +16,14 @@ data class NhtsaDecodeVinDto(
         @SerialName("Value") val value: String? = null,
         @SerialName("ValueId") val valueId: String? = null,
         @SerialName("Variable") val variable: String? = null,
-        @SerialName("VariableId") val variableId: Int? = null
+        @SerialName("VariableId") val variableId: Long? = null
     )
 
     companion object {
-        const val ERROR_TEXT_VARIABLE_ID = 191
-        const val MAKE_VARIABLE_ID = 26
-        const val MODEL_VARIABLE_ID = 28
-        const val VEHICLE_TYPE_VARIABLE_ID = 39
-        const val BODY_CLASS_VARIABLE_ID = 5
+        const val ERROR_TEXT_VARIABLE_ID = 191L
+        const val MAKE_VARIABLE_ID = 26L
+        const val MODEL_VARIABLE_ID = 28L
+        const val VEHICLE_TYPE_VARIABLE_ID = 39L
+        const val BODY_CLASS_VARIABLE_ID = 5L
     }
 }
